requirements.txt
colorama
python-dotenv
requests
aiohttp
asyncio